requirements.txt
streamlit
Markdown
anthropic
async-timeout
python-dotenv
numpy==1.26.4
transformers==4.41.0
llama-cloud==0.1.26
llama-cloud-services==0.6.31
llama-index==0.12.41
llama-index-agent-openai==0.4.10
llama-index-cli==0.4.3
llama-index-core==0.12.41
llama-index-embeddings-huggingface==0.5.4
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.6
llama-index-llms-openai==0.4.5
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-index-retrievers-bm25==0.5.2
llama-index-vector-stores-qdrant
llama-parse==0.6.31
qdrant-client
snowflake-connector-python
matplotlib
kneed